refactor: step2 패키지 구성 변경 및 getter 삭제 및 toString으로 대체
