+++ src/main/java/persistence/sql/ddl/usecase/GetFieldFromClassUseCase.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.usecase;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.GeneratedValue;
