+++ src/main/java/persistence/sql/ddl/usecase/GetTableNameFromClassUseCase.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.usecase;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
