+++ src/main/java/persistence/sql/ddl/vo/TableName.java
@@ -13,10 +13,6 @@ public class TableName {
         return new TableName(name);
     }
 
-    public String getName() {
-        return name;
-    }
-
     @Override
     public String toString() {
         return name;
